Pass map markers as props and solve wrong types on graphql schema

+++ api/src/graphql/companies.sdl.ts
@@ -2,8 +2,8 @@ export const schema = gql`
   type Company {
     id: String!
     name: String!
-    latitude: Int!
-    longitude: Int!
+    latitude: Float!
+    longitude: Float!
     Developer: [Developer]!
     Office: [Office]!
   }
@@ -15,14 +15,14 @@ export const schema = gql`
 
   input CreateCompanyInput {
     name: String!
-    latitude: Int!
-    longitude: Int!
+    latitude: Float!
+    longitude: Float!
   }
 
   input UpdateCompanyInput {
     name: String
-    latitude: Int
-    longitude: Int
+    latitude: Float
+    longitude: Float
   }
 
   type Mutation {
